agents/2_inputs/requirements.txt
azure-ai-projects
python-dotenv
azure-identity
pytest